\name{data.x}
\alias{data.x}
\docType{data}
\title{data.x}

\description{Vector indicating disease status for each subject. The vector data.x 
is a binary vector of length \emph{n}, where \emph{n} is the number of samples.}
\format{
 int [1:\emph{n}] 0 0 1 0 0 1 1 0 1 0 ...
}
\keyword{ datasets }