\name{data.y}
\alias{data.y}
\docType{data}
\title{An array of gene expression data.}
\description{
A 3-dimensional array containing probe-level gene expression data.  The array data.y contains values for \emph{S} probesets, \emph{n} samples, and \emph{P} probes.
}

\format{
The format is:
 num [1:\emph{S}, 1:\emph{n}, 1:\emph{P}] 10.03 6.09 7.27 9.85 7.47 ...
}
\keyword{ datasets }