\name{EquiNorm-internal}

\alias{cel.names}
\alias{chip.id}
\alias{chip.ss}
\alias{ctrl.name}
\alias{data.code}
\alias{folder.name}
\alias{G}
\alias{gene.name}
\alias{i}
\alias{idx.ctrl}
\alias{m}
\alias{P}
\alias{probe.name}
\alias{probe.table}
\alias{temp}
\alias{work.path}

\alias{affy.norm}
\alias{affy.norm.EM}
\alias{affy.norm.Estep}
\alias{affy.norm.Mstep}
\alias{affy.norm.start}
\alias{compute.dens.D}
\alias{compute.dens.ND}
\alias{compute.resid.ctd}
\alias{equi.gene.norm.EM}
\alias{equi.gene.norm.Estep}
\alias{equi.gene.norm.Mstep}
\alias{equi.gene.norm.start }
\alias{gene.select.fast}
\alias{knn.cv.one}
\alias{lda.cv.one}
\alias{lps.cv.one}
\alias{new.knn}
\alias{new.t.test}
\alias{new.t.test.fast}
\alias{roc.emp}
\alias{sample2}
\alias{train.classifier}
\alias{train.cv}
\alias{train.cv.one}
\alias{train.knn.easy}
\alias{train.pam}
\alias{train.pam.bkup}

\title{Internal EquiNorm Functions}
\description{
 Internal EquiNorm functions
}
\details{
These are not to be called by the user.
}
\keyword{ internal }


